+++ src/Functions/exports/CreateApplicationInsightsProject.ps1
@@ -30,7 +30,7 @@ PS C:\> {{ Add code here }}
 .Outputs
 System.Object
 .Link
-https://docs.microsoft.com/en-us/powershell/module/az.functions/createapplicationinsightsproject
+https://docs.microsoft.com/powershell/module/az.functions/createapplicationinsightsproject
 #>
 function CreateApplicationInsightsProject {
 [CmdletBinding(PositionalBinding=$false)]