+++ src/Websites/Websites/Cmdlets/DeploymentSlots/SetAzureWebAppSlot.cs
@@ -181,7 +181,8 @@ namespace Microsoft.Azure.Commands.WebApps.Cmdlets.DeploymentSlots
                                 parameters.Contains("Use32BitWorkerProcess") ? (bool?)Use32BitWorkerProcess : null,
                             AutoSwapSlotName = parameters.Contains("AutoSwapSlotName") ? AutoSwapSlotName : null,
                             NumberOfWorkers = parameters.Contains("NumberOfWorkers") ? NumberOfWorkers : WebApp.SiteConfig.NumberOfWorkers,
-                            AlwaysOn = parameters.Contains("AlwaysOn") ? (bool)AlwaysOn : false
+                            AlwaysOn = parameters.Contains("AlwaysOn") ? (bool?)AlwaysOn : null,
+                            FtpsState = parameters.Contains("FtpsState") ? FtpsState : WebApp.SiteConfig.FtpsState
                         };
                     }