+++ src/DesktopVirtualization/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 3.0.0
 * Upgraded api version to 2021-02-01-preview.
 
 ## Version 2.1.1