+++ src/StackHCI/ChangeLog.md
@@ -19,6 +19,10 @@
 -->
 ## Upcoming Release
 
+## Version 0.7.0
+* Made changes in the registration to onboard nodes to Azure Arc.
+    - Registering On-Premises Azure Stack HCI with Azure will also make the nodes in the cluster Azure Arc enabled.   
+
 ## Version 0.6.0
 * Fixed addition of assigned roles list.